pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mgmtlit"
version = "0.1.0"
description = "AI-assisted literature review pipeline for management, organizations, and information systems"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "MgmtLit Contributors"}]
dependencies = [
  "httpx>=0.27.2",
  "pydantic>=2.9.2",
  "typer>=0.12.5",
  "rich>=13.9.2",
  "python-dotenv>=1.0.1",
  "tenacity>=9.0.0",
  "openai>=1.52.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.9",
]

[project.scripts]
mgmtlit = "mgmtlit.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
